Live Activity for current track
