+++ NowPlaying/NowPlayingBundle.swift
@@ -1,0 +1,11 @@
+// Made by Lumaa
+
+import WidgetKit
+import SwiftUI
+
+@main
+struct NowPlayingBundle: WidgetBundle {
+    var body: some Widget {
+        NowPlayingLiveActivity()
+    }
+}
